pyproject.toml
[tool.poetry]
name = "board"
version = "0.1.0"
description = ""
authors = ["Jueuunn7 <wndmsrla1234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
django-ninja = "^1.3.0"
django-ninja-extra = "^0.22.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
